assignments/requirements.txt
# requirements.txt

torch>=2.0.0
transformers>=4.0.0
datasets>=2.0.0
PyYAML>=6.0
tqdm>=4.0.0
numpy>=1.21.0